refactor(kfx): separate runtime warrant authority
