+++ developer/sdk/kfd/support-matrix.json
@@ -532,7 +532,7 @@
       "implementation": {
         "status": "implemented-specialized-witness",
         "surfaces": [
-          "framework/core/src/libkungfu/src/runtime/kfx/native_registry.cpp",
+          "framework/core/src/libkungfu/src/runtime/kfx/native_authority.cpp",
           "framework/kfx/kungfu-kfx-domain-profile.contract.json",
           "framework/core/src/python/kungfu/storage/kfx_service.py"
         ]
@@ -542,7 +542,7 @@
         "evidenceRoots": [
           {
             "path": "framework/kfx/evidence/kfd-10/runtime-warrant-adopter.json",
-            "sha256": "sha256:68a7cb0b50e5f0f4d0aeeac5502d8cfcd1f50bcfd6e1ee5e6770356e84adecde"
+            "sha256": "sha256:b468cbefe4ab497165128557d3dd118dedf50fd69630e09306b94c2d5fd5fa3c"
           }
         ]
       },
