+++ scripts/kfd-support-matrix.test.mjs
@@ -110,7 +110,7 @@ test('validates the exact KFD-1 through KFD-13 authority', (t) => {
   assert.equal(kfd10.supportStatus, 'draft-adopter-evidence');
   assert.equal(kfd10.implementation.status, 'implemented-specialized-witness');
   assert.deepEqual(kfd10.implementation.surfaces, [
-    'framework/core/src/libkungfu/src/runtime/kfx/native_registry.cpp',
+    'framework/core/src/libkungfu/src/runtime/kfx/native_authority.cpp',
     'framework/kfx/kungfu-kfx-domain-profile.contract.json',
     'framework/core/src/python/kungfu/storage/kfx_service.py',
   ]);
